+++ R/data_precooked.R
@@ -43,3 +43,9 @@
 #' @rdname made_earlier
 #' @usage data(res_siena_selection)
 "res_siena_selection"
+
+#' @docType data
+#' @keywords datasets
+#' @rdname made_earlier
+#' @usage data(res_monan_traces)
+"res_monan_traces"